+++ package/src/presets/react/index.ts
@@ -1,3 +1,5 @@
+export type { ReactPresetOptions } from "#/presets/react/preset";
+
 export { PLUGIN_REACT } from "#/presets/react/plugin";
 export { reactPreset } from "#/presets/react/preset";
 export {
@@ -9,3 +11,7 @@ export {
     RULES_REACT_STYLE,
     RULES_REACT_SUSPICIOUS,
 } from "#/presets/react/rules/react";
+export {
+    RULES_REACT_DEBUG,
+    RULES_REACT_DEBUG_RESTRICTION,
+} from "#/presets/react/rules/react-debug";