+++ package/src/presets/react/preset/index.ts
@@ -2,8 +2,15 @@ import type { Preset, PresetResult } from "#/@types/preset";
 
 import { PLUGIN_REACT } from "#/presets/react/plugin";
 import { RULES_REACT } from "#/presets/react/rules/react";
+import { RULES_REACT_DEBUG } from "#/presets/react/rules/react-debug";
 
-const reactPreset = (): Preset => {
+type CompleteReactPresetOptions = {
+    debug: boolean;
+};
+
+type ReactPresetOptions = Partial<CompleteReactPresetOptions>;
+
+const reactPreset = (options?: ReactPresetOptions): Preset => {
     return (): PresetResult => {
         return {
             config: {
@@ -12,10 +19,12 @@ const reactPreset = (): Preset => {
                 ],
                 rules: {
                     ...RULES_REACT,
+                    ...(options?.debug === true ? RULES_REACT_DEBUG : {}),
                 },
             },
         };
     };
 };
 
+export type { CompleteReactPresetOptions, ReactPresetOptions };
 export { reactPreset };